pyproject.toml
[project]
name = "vtk-image-labeler-3d"
version = "0.1.0"
description = ""
authors = [
    { name = "Jinkoo", email = "jinkoo2@hotmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi == 2024.12.14",
    "charset-normalizer == 3.4.1",
    "contourpy == 1.3.1",
    "cycler == 0.12.1",
    "fonttools == 4.55.3",
    "idna == 3.10",
    "kiwisolver == 1.4.8",
    "matplotlib == 3.10.0",
    "numpy == 2.1.3",
    "opencv-python == 4.10.0.84",
    "packaging == 24.2",
    "pillow == 11.0.0",
    "platformdirs == 4.3.6",
    "pooch == 1.8.2",
    "pydicom == 3.0.1",
    "pyparsing == 3.2.0",
    "PyQt5 == 5.15.11",
    "PyQt5-Qt5 == 5.15.2",
    "PyQt5_sip == 12.16.1",
    "python-dateutil == 2.9.0.post0",
    "requests == 2.32.3",
    "scooby == 0.10.0",
    "SimpleITK == 2.4.0",
    "six == 1.17.0",
    "typing_extensions == 4.12.2",
    "urllib3 == 2.3.0",
    "vtk == 9.3.1",
    "python-dotenv",  # No version specified, Poetry will resolve
    "scikit-image (>=0.25.2,<0.26.0)"
]

[tool.poetry.scripts]
main = "vtk_image_labeler_3d.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"